Difficulty and Dumbasses update
Now there are different difficulties with the ratios of different Market Strategies the bots will globally have.
Also now there is a new strategy implemented, the dumb ass. He only buys bad assets or random one if he can't find any and sells only if he made a profit.
Also now using apache commons the number of bots is modifiable by entering it as an argument, and also a difficulty can be entered.
If no difficulty is entered the strategies of the bots will be equally distributed among the implemented ones (random and dumbass).