Streamline Update
Now the eventHandler contains the game and runs everything that happens on it, since this is essential so that the market refreshes every 6 hours so we can have a solid market we can build the trading around of. Also some minor bugs have been fixed due to this change, like bots doing things after they have been hospitalized because the execution time of their update events was after the asynchronous flushing of dead bots. Also the same thing used to happen with bots selling assets that asynchronously had been flushed due to bankruptcy if they had died and wanted to sell a bankrupt asset which would not have been cleared since they were dead, which raised an exception.

Also now levels have a predetermined number of bots so that if you don't enter a parameter of bots (I think this can be a "custom play" mode), the game can have a number of bots that will make itself interesting and suit the difficulty level.
Also so as to not keep the irrelevant information and as a start point for the rest of the market strategies, the strategies now have a clear memory method.