Strategies Update
Now all the strategies have been added, in the future I will probably polish them up for code cleanness.

Also now the assets' industryBooms and industryCrashes will last a few turns so that our bots' strategies work better.

In the future I think we should consider adding in the analysing the state of the market action for the player maybe giving a tip on wether some asset's industry is booming.

Either way this version could be considered a beta since all the game difficulties (not that we've given them much thought) are playable as of now.