STRATEGY PATTERN for Bots to decide what action to do
Currently they just have either the MoveFastStrategy (that has a high prob of moving elsewhere) or the SedentaryStrategy (low prob).
There are two more strategies implemented: seekFood and seekMental.
The idea is that later on the bots will have states,  that will determine their chooseActionStrategies and their socialStrategies. For example if a bot is getting hungry, he'll enter the hungry state in which he's actionStrategy would be seekFood and his socialStrategy beRough.