BotStates update
Now bots have states which using the states pattern govern their actions.

This is the core of their AI and it will make them respond to game stimuli like low mental health and hunger.

Using the BotStateGovernor they will change this state according to the stimuli.

Their ability to adapt to stimuli is global for all bots in the difficulty mode you play.